feat: implement Order management feature
- Create Order and OrderItem entities with 1-to-N relationship and Cascade.
- Add product unitPrice snapshot to OrderItem.
- Implement OrderController for full CRUD and related service logic.
- Create DTOs (OrderRequest, OrderResponse, etc.) for requests and responses.
- Compute total price based on line items' snapshot price dynamically.
- Implement GlobalExceptionHandler to handle generic exceptions elegantly.
- Refactor unused OrderItemRepository and unused imports.
- Update project documentation in README.

+++ src/main/java/com/example/demo1/dto/ProductResponse.java
@@ -3,6 +3,8 @@ package com.example.demo1.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import com.example.demo1.model.Product;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Product representation returned to the client")
 public record ProductResponse(
 
@@ -12,17 +14,17 @@ public record ProductResponse(
 
         @Schema(description = "Product description", example = "Laptop 14 inch, 16GB RAM") String description,
 
-        @Schema(description = "Product price", example = "15000000") Double price,
+        @Schema(description = "Product price", example = "15000000") BigDecimal price,
 
         @Schema(description = "Product stock", example = "10") Integer stock) {
 
-    /** Maps an entity to a response DTO. */
+    /** Maps an entity to a response DTO, avoiding floating-point scientific notation. */
     public static ProductResponse from(Product product) {
         return new ProductResponse(
                 product.getId(),
                 product.getName(),
                 product.getDescription(),
-                product.getPrice(),
+                BigDecimal.valueOf(product.getPrice()),
                 product.getStock());
     }
 }
